+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+### Changed
+
+- Added support for dictionaries of nested models
+- Added support for deeply nested models beyond level-1 deep including:
+  - dictionaries of lists of ... of nested models
+  - lists of tuples of lists .... of nested models
+
+### Fixed
+
+- Fixed `AttributeError: 'get_primary_key_field'` when None is passed to a field with an optional nested model
+
+
 ## [0.6.0] - 2024-07-01
 
 ### Added